Moved fixtures to conftest, for reuse across test suites & files.
