+++ requirements.txt
@@ -1,4 +1,5 @@
 -e git://github.com/Shizmob/pydle.git@asyncio#egg=pydle
 pytest
 pytest-cov
+pytest-asyncio
 aiounittest
